jwt fix: add secure flag to coockie beacuse if same site  be none coockie must be secure adn cors be defiend with exact origins then request send from client with withCredential true

+++ src/prisma/prisma.service.ts
@@ -1,9 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, OnModuleDestroy, OnModuleInit } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { PrismaClient } from '@prisma/client';
 
 @Injectable()
-export class PrismaService extends PrismaClient {
+export class PrismaService
+  extends PrismaClient
+  implements OnModuleInit, OnModuleDestroy
+{
   constructor(config: ConfigService) {
     super({
       datasources: {
@@ -14,7 +17,19 @@ export class PrismaService extends PrismaClient {
     });
   }
 
-  cleanDb() {
+  async onModuleInit() {
+    await this.$connect();
+  }
+
+  async onModuleDestroy() {
+    await this.$disconnect();
+  }
+
+  async cleanDb() {
+    if (process.env.NODE_ENV === 'production') return;
+    // const models = Reflect.ownKeys(this).filter((key) => key[0] !== '_');
+    // return Promise.all(models.map((modelKey) => this[modelKey].deleteMany()));
+
     return this.$transaction([
       this.order.deleteMany(),
       this.token.deleteMany(),
